KIRAGAG

+++ app/Http/Controllers/PostController.php
@@ -85,6 +85,13 @@ class PostController extends Controller
     public function edit($id)
     {
         $post= Post::find($id);
+
+        // check for correct user
+
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unathorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -120,6 +127,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post= Post::find($id);
+
+        // check for correct user
+
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unathorized Page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
